// src/cli/types.ts
/** Command-line options accepted by the CLI. */
export interface CommandLineOptions {
    /** Specific number of random quotes to display. */
    number?: number;
  
    /** Random quote from the specified character to display. */
    character?: string;
  
    /** Random quote containing the keyword to display. */
    keyword?: string;
  }
  